New 'one form per object' option.
By default, Datatrans displays all objects for a model on one page, in
the same HTML form. This leads to ridiculous load times for any database
worth speaking of.

This commit adds an option, 'one_form_per_object', which you can set in
translation descriptor classes. For models where this is set, the
translation form is split and you get one per object for that model.